% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitNB}
\alias{fitNB}
\title{Fit Negative Binomial Models to Gene Expression Data}
\usage{
fitNB(data, assay_type = "counts")
}
\arguments{
\item{data}{A `SpatialExperiment` object containing gene expression data.}

\item{assay_type}{The type of assay to use from the SpatialExperiment object,
with the default being 'counts'.}
}
\value{
A list of `glm.nb` model objects for each gene, where the model fitting
was successful.
}
\description{
Fits Negative Binomial models to gene expression data contained in a
`SpatialExperiment` object. This function iterates over each gene, fits a model
using gene expression counts, and stores the resulting model if successful.
}
\examples{
# Assuming `top_500` is a preprocessed SpatialExperiment object
nb_models <- fitNB(top_500)

}
