% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{modelFit}
\alias{modelFit}
\title{Comprehensive Analysis Pipeline for SpatialExperiment Data}
\usage{
modelFit(seo, n)
}
\arguments{
\item{seo}{A `SpatialExperiment` object to be processed.}

\item{n}{The number of top variable genes to select for modeling.}
}
\value{
A list of `glm.nb` model objects for each gene where the model fitting
was successful.
}
\description{
This function provides an end-to-end analysis pipeline for `SpatialExperiment` objects.
It first normalizes the data using library size normalization, then selects the top
variable genes, and finally fits Negative Binomial models to the gene expression data.
This pipeline is designed to be a convenient wrapper for a series of complex steps,
streamlining the analysis of spatial transcriptomics data.
}
\examples{
# Assuming 'seo' is a preprocessed SpatialExperiment object
models <- modelFit(seo, n = 500)

}
