% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{select_features}
\alias{select_features}
\title{Select Top Variable Genes from a SpatialExperiment Object}
\usage{
select_features(SEO, n, assay_name = "logcounts")
}
\arguments{
\item{SEO}{A `SpatialExperiment` object to analyze.}

\item{n}{The number of top variable genes to select.}

\item{assay_name}{The name of the assay to use for variability calculations.}
}
\value{
A subset of the `SpatialExperiment` object containing only the top `n`
most variable genes.
}
\description{
Select Top Variable Genes from a SpatialExperiment Object
}
\examples{
# Assuming 'SEO' is a SpatialExperiment object with logcounts
SEO_hvg = select_features(SEO, n = 500, assay_name = "logcounts")

}
